requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
httpx==0.25.0
python-dotenv==1.0.0
structlog==23.1.0
slowapi==0.1.8 # For rate limiting